src/plugins/status/requirements.txt
nonebot2>=2.0.0b5
nonebot-adapter-onebot>=2.2.3
pillow>=9.1.1
py-cpuinfo
psutil
httpx